refactor: consolidate duplicated ViewModelToViewConverter into shared PropertyEditorContextToViewConverter
