+++ src/Beutl.Editor.Components/ColorGradingProperties/Views/ColorGradingPropertiesEditor.axaml.cs
@@ -1,10 +1,8 @@
-﻿using Avalonia;
-using Avalonia.Animation;
+﻿using Avalonia.Animation;
 using Avalonia.Controls;
 using Avalonia.Controls.Primitives;
-using Avalonia.Data;
-using Avalonia.Data.Converters;
 using Avalonia.Interactivity;
+using Beutl.Controls.Converters;
 using Beutl.Editor.Components.ColorGradingProperties.ViewModels;
 using Beutl.Editor.Components.ColorGradingTab.ViewModels;
 using Microsoft.Extensions.DependencyInjection;
@@ -18,7 +16,7 @@ public sealed partial class ColorGradingPropertiesEditor : UserControl
 
     public ColorGradingPropertiesEditor()
     {
-        Resources["ViewModelToViewConverter"] = ViewModelToViewConverterImpl.Instance;
+        Resources["ViewModelToViewConverter"] = PropertyEditorContextToViewConverter.Instance;
         InitializeComponent();
 
         expandProps.GetObservable(ToggleButton.IsCheckedProperty)
@@ -54,37 +52,4 @@ public sealed partial class ColorGradingPropertiesEditor : UserControl
         }
     }
 
-    private sealed class ViewModelToViewConverterImpl : IValueConverter
-    {
-        public static readonly ViewModelToViewConverterImpl Instance = new();
-
-        public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
-        {
-            if (value is IPropertyEditorContext viewModel)
-            {
-                if (viewModel.Extension.TryCreateControl(viewModel, out var control))
-                {
-                    return control;
-                }
-                else
-                {
-                    return new Label
-                    {
-                        Height = 24,
-                        Margin = new Thickness(0, 4),
-                        Content = viewModel.Extension.DisplayName
-                    };
-                }
-            }
-            else
-            {
-                return BindingNotification.Null;
-            }
-        }
-
-        public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
-        {
-            return BindingNotification.Null;
-        }
-    }
 }
