+++ src/Beutl/Views/Editors/GraphModelNodeMemberView.axaml.cs
@@ -1,10 +1,8 @@
-﻿using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Data;
-using Avalonia.Data.Converters;
+﻿using Avalonia.Controls;
 using Avalonia.Interactivity;
 
 using Beutl.Controls;
+using Beutl.Controls.Converters;
 using Beutl.ViewModels.Editors;
 
 namespace Beutl.Views.Editors;
@@ -13,7 +11,7 @@ public partial class GraphModelNodeMemberView : UserControl
 {
     public GraphModelNodeMemberView()
     {
-        Resources["ViewModelToViewConverter"] = ViewModelToViewConverter.Instance;
+        Resources["ViewModelToViewConverter"] = PropertyEditorContextToViewConverter.Instance;
         InitializeComponent();
         ExpandTransitionHelper.Attach(expandToggle, content, ExpandTransitionHelper.ListItemDuration);
     }
@@ -51,37 +49,4 @@ public partial class GraphModelNodeMemberView : UserControl
         }
     }
 
-    private sealed class ViewModelToViewConverter : IValueConverter
-    {
-        public static readonly ViewModelToViewConverter Instance = new();
-
-        public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
-        {
-            if (value is IPropertyEditorContext viewModel)
-            {
-                if (viewModel.Extension.TryCreateControl(viewModel, out var control))
-                {
-                    return control;
-                }
-                else
-                {
-                    return new Label
-                    {
-                        Height = 24,
-                        Margin = new Thickness(0, 4),
-                        Content = viewModel.Extension.DisplayName
-                    };
-                }
-            }
-            else
-            {
-                return BindingNotification.Null;
-            }
-        }
-
-        public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
-        {
-            return BindingNotification.Null;
-        }
-    }
 }
