+++ src/Beutl/Views/Editors/PropertiesEditor.axaml.cs
@@ -1,7 +1,6 @@
-﻿using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Data;
-using Avalonia.Data.Converters;
+﻿using Avalonia.Controls;
+
+using Beutl.Controls.Converters;
 
 namespace Beutl.Views.Editors;
 
@@ -9,41 +8,7 @@ public partial class PropertiesEditor : UserControl
 {
     public PropertiesEditor()
     {
-        Resources["ViewModelToViewConverter"] = ViewModelToViewConverter.Instance;
+        Resources["ViewModelToViewConverter"] = PropertyEditorContextToViewConverter.Instance;
         InitializeComponent();
-    }
-
-    public sealed class ViewModelToViewConverter : IValueConverter
-    {
-        public static readonly ViewModelToViewConverter Instance = new();
-
-        public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
-        {
-            if (value is IPropertyEditorContext viewModel)
-            {
-                if (viewModel.Extension.TryCreateControl(viewModel, out var control))
-                {
-                    return control;
-                }
-                else
-                {
-                    return new Label
-                    {
-                        Height = 24,
-                        Margin = new Thickness(0, 4),
-                        Content = viewModel.Extension.DisplayName
-                    };
-                }
-            }
-            else
-            {
-                return BindingNotification.Null;
-            }
-        }
-
-        public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
-        {
-            return BindingNotification.Null;
-        }
     }
 }
